requirement.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.3.0
certifi==2025.4.26
click==8.2.0
confluent-kafka==2.10.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.104.1
h11==0.16.0
httpcore==1.0.9
httpx==0.25.1
idna==3.10
iniconfig==2.1.0
kafka-python==2.2.7
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pip==24.3.1
pluggy==1.6.0
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
pydantic-settings==2.9.1
PyJWT==2.10.1
pytest==7.4.3
python-dotenv==1.0.0
redis==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.27.0
typing_extensions==4.13.2
typing-inspection==0.4.0
uvicorn==0.23.2